@@ -3,21 +3,21 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@ant-design/icons": "^4.6.2",
-    "@material-ui/core": "^4.11.3",
-    "@testing-library/jest-dom": "^5.11.10",
-    "@testing-library/react": "^11.2.6",
-    "@testing-library/user-event": "^12.8.3",
-    "antd": "^4.15.0",
-    "axios": "^0.21.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "react-scrollspy-nav": "^1.2.4",
-    "react-slick": "^0.28.1",
+    "@ant-design/icons": "^4.7.0",
+    "@material-ui/core": "^4.12.4",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^14.2.1",
+    "antd": "^4.21.6",
+    "axios": "^0.27.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
+    "react-scrollspy-nav": "^1.3.0",
+    "react-slick": "^0.29.0",
     "slick-carousel": "^1.8.1",
-    "web-vitals": "^1.1.1"
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "PORT=2929 react-scripts start",    
