@@ -27,40 +27,40 @@ import SF5 from "../../images/4ssdd5.png";
 import SF6 from "../../images/4ssdd6.png";
 
 // aeloop brokers
-import A1 from "../../images/a1.png";
-import A2 from "../../images/a2.png";
-import A3 from "../../images/a3.png";
-import A4 from "../../images/a4.png";
-import A5 from "../../images/a5.png";
-import A6 from "../../images/a6.png";
-import A7 from "../../images/a7.png";
-import A8 from "../../images/a8.png";
-import A9 from "../../images/a9.png";
-import A10 from "../../images/a10.png";
-import A11 from "../../images/a11.png";
-import A12 from "../../images/a12.png";
-import A13 from "../../images/a13.png";
-import A14 from "../../images/a14.png";
-import A15 from "../../images/a15.png";
-import A16 from "../../images/a16.png";
-import A17 from "../../images/a17.png";
-import A18 from "../../images/a18.png";
-import A19 from "../../images/a19.png";
-import A20 from "../../images/a20.png";
-import A21 from "../../images/a21.png";
-import A22 from "../../images/a22.png";
-import A23 from "../../images/a23.png";
-import A24 from "../../images/a24.png";
-import A25 from "../../images/a25.png";
-import A26 from "../../images/a26.png";
-import A27 from "../../images/a27.png";
-import A28 from "../../images/a28.png";
-import A29 from "../../images/a29.png";
-import A30 from "../../images/a30.png";
-import A31 from "../../images/a31.png";
-import A32 from "../../images/a32.png";
-import A33 from "../../images/a33.png";
-import A34 from "../../images/a34.png";
+// import A1 from "../../images/a1.png";
+// import A2 from "../../images/a2.png";
+// import A3 from "../../images/a3.png";
+// import A4 from "../../images/a4.png";
+// import A5 from "../../images/a5.png";
+// import A6 from "../../images/a6.png";
+// import A7 from "../../images/a7.png";
+// import A8 from "../../images/a8.png";
+// import A9 from "../../images/a9.png";
+// import A10 from "../../images/a10.png";
+// import A11 from "../../images/a11.png";
+// import A12 from "../../images/a12.png";
+// import A13 from "../../images/a13.png";
+// import A14 from "../../images/a14.png";
+// import A15 from "../../images/a15.png";
+// import A16 from "../../images/a16.png";
+// import A17 from "../../images/a17.png";
+// import A18 from "../../images/a18.png";
+// import A19 from "../../images/a19.png";
+// import A20 from "../../images/a20.png";
+// import A21 from "../../images/a21.png";
+// import A22 from "../../images/a22.png";
+// import A23 from "../../images/a23.png";
+// import A24 from "../../images/a24.png";
+// import A25 from "../../images/a25.png";
+// import A26 from "../../images/a26.png";
+// import A27 from "../../images/a27.png";
+// import A28 from "../../images/a28.png";
+// import A29 from "../../images/a29.png";
+// import A30 from "../../images/a30.png";
+// import A31 from "../../images/a31.png";
+// import A32 from "../../images/a32.png";
+// import A33 from "../../images/a33.png";
+// import A34 from "../../images/a34.png";
 
 // sparkfost-vax
 import SparkVax from "../../images/sparksoftvax.png";
